# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Omni Writer API
  description: API para geração de artigos com IA
  version: 2.0.0
  contact:
    name: Omni Writer Team
    email: support@omniwriter.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.omniwriter.com
    description: Production server
paths:
...
  /api/blogs:
    get:
      summary: Listar blogs
      description: Retorna lista de blogs disponíveis
      tags:
        - Blogs
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de blogs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Blog'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Criar blog
      description: Cria um novo blog
      tags:
        - Blogs
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlogCreate'
      responses:
        '201':
          description: Blog criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Blog'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/blogs/{blog_id}:
    delete:
      summary: Deletar blog
      description: Remove um blog específico
      tags:
        - Blogs
      security:
        - BearerAuth: []
      parameters:
        - name: blog_id
          in: path
          required: true
          schema:
            type: integer
          description: ID do blog a ser deletado
      responses:
        '200':
          description: Blog deletado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Blog deletado com sucesso"
        '404':
          description: Blog não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/blogs/{blog_id}/prompts:
    get:
      summary: Listar prompts do blog
      description: Retorna lista de prompts de um blog específico
      tags:
        - Prompts
      security:
        - BearerAuth: []
      parameters:
        - name: blog_id
          in: path
          required: true
          schema:
            type: integer
          description: ID do blog
      responses:
        '200':
          description: Lista de prompts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Prompt'
        '404':
          description: Blog não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Criar prompt
      description: Cria um novo prompt para um blog
      tags:
        - Prompts
      security:
        - BearerAuth: []
      parameters:
        - name: blog_id
          in: path
          required: true
          schema:
            type: integer
          description: ID do blog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromptCreate'
      responses:
        '201':
          description: Prompt criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prompt'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Blog não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/blogs/{blog_id}/prompts/{prompt_id}:
    delete:
      summary: Deletar prompt
      description: Remove um prompt específico de um blog
      tags:
        - Prompts
      security:
        - BearerAuth: []
      parameters:
        - name: blog_id
          in: path
          required: true
          schema:
            type: integer
          description: ID do blog
        - name: prompt_id
          in: path
          required: true
          schema:
            type: integer
          description: ID do prompt
      responses:
        '200':
          description: Prompt deletado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Prompt deletado com sucesso"
        '404':
          description: Blog ou prompt não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT para autenticação
  schemas:
    Blog:
      type: object
      properties:
        id:
          type: integer
          description: ID único do blog
          example: 1
        nome:
          type: string
          maxLength: 40
          description: Nome do blog
          example: "Tech Blog"
        desc:
          type: string
          maxLength: 80
          nullable: true
          description: Descrição do blog
          example: "Blog sobre tecnologia e inovação"
      required:
        - id
        - nome
    BlogCreate:
      type: object
      properties:
        nome:
          type: string
          maxLength: 40
          description: Nome do blog
          example: "Tech Blog"
        desc:
          type: string
          maxLength: 80
          nullable: true
          description: Descrição do blog
          example: "Blog sobre tecnologia e inovação"
      required:
        - nome
    Prompt:
      type: object
      properties:
        id:
          type: integer
          description: ID único do prompt
          example: 1
        blog_id:
          type: integer
          description: ID do blog associado
          example: 1
        titulo:
          type: string
          maxLength: 100
          description: Título do prompt
          example: "Como implementar IA"
        conteudo:
          type: string
          maxLength: 2000
          description: Conteúdo do prompt
          example: "Escreva um artigo sobre implementação de IA..."
        ativo:
          type: boolean
          description: Status do prompt
          example: true
      required:
        - id
        - blog_id
        - titulo
        - conteudo
        - ativo
    PromptCreate:
      type: object
      properties:
        titulo:
          type: string
          maxLength: 100
          description: Título do prompt
          example: "Como implementar IA"
        conteudo:
          type: string
          maxLength: 2000
          description: Conteúdo do prompt
          example: "Escreva um artigo sobre implementação de IA..."
        ativo:
          type: boolean
          default: true
          description: Status do prompt
          example: true
      required:
        - titulo
        - conteudo
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Código do erro
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Mensagem de erro
          example: "Dados inválidos fornecidos"
        details:
          type: object
          nullable: true
          description: Detalhes do erro
          example:
            field: "nome"
            constraint: "max_length"
        trace_id:
          type: string
          description: ID de rastreamento
          example: "err_20250127_001"
        timestamp:
          type: string
          format: date-time
          description: Timestamp do erro
          example: "2025-01-27T21:30:00Z"
      required:
        - error
        - message
        - timestamp
...